Поделила на модули
